redefine of MCP modes + improvements on the chatId gestion in tools
