+++ apps/backend/src/mcp/server.ts
@@ -4,10 +4,9 @@ import { StreamableHTTPServerTransport } from '@modelcontextprotocol/sdk/server/
 import { listUserProjects } from '../queries/project.queries';
 import type { McpEndpointSettings } from '../types/mcp-endpoint';
 import { registerNaoMcpApps } from './embed/ui-resources';
-import { registerAgentTools } from './tools/agent';
-import { registerDataTools } from './tools/data';
-import { registerFileTools } from './tools/files';
-import { registerStoryTools } from './tools/stories';
+import { registerContextLayerTools } from './tools/context-layer';
+import { registerStoryManagementTools } from './tools/story-management';
+import { registerSubAgentTools } from './tools/sub-agent';
 
 export interface McpSession {
 	transport: StreamableHTTPServerTransport;
@@ -49,23 +48,17 @@ export async function resolveProjectId(userId: string): Promise<string> {
 
 export function createMcpServer(userId: string, projectId: string, settings: McpEndpointSettings): McpServer {
 	const server = new McpServer({ name: 'nao', version: '0.1.0' }, { capabilities: { tools: {}, resources: {} } });
-	const sessionChatRef: { lastChatId?: string } = {};
-	const ctx = { userId, projectId, settings, sessionChatRef };
+	const ctx = { userId, projectId, settings };
 
-	if (settings.agentModeEnabled) {
-		registerAgentTools(server, ctx);
+	if (settings.subAgentModeEnabled) {
+		registerSubAgentTools(server, ctx);
 	}
-	if (settings.toolsModeEnabled) {
-		registerDataTools(server, ctx);
-		registerFileTools(server, ctx);
-	}
-	if (settings.objectsModeEnabled) {
-		registerStoryTools(server, ctx);
+	if (settings.contextLayerModeEnabled) {
+		registerContextLayerTools(server, ctx);
 	}
 
-	if (settings.toolsModeEnabled || settings.objectsModeEnabled) {
-		registerNaoMcpApps(server);
-	}
+	registerStoryManagementTools(server, ctx);
+	registerNaoMcpApps(server);
 
 	return server;
 }
