+++ apps/backend/src/mcp/tools/sub-agent.ts
@@ -1,4 +1,5 @@
 import type { McpServer } from '@modelcontextprotocol/sdk/server/mcp.js';
+import { buildStoryChartBlock } from '@nao/shared';
 import { type InferUIMessageChunk, readUIMessageStream } from 'ai';
 import { z } from 'zod';
 
@@ -12,32 +13,62 @@ import { defineMcpHandler } from '../logging';
 import { chatUrl } from '../urls';
 
 const ASK_NAO_DESCRIPTION =
-	'Ask nao an analytics question in natural language. Creates a chat that is visible in the UI.\n\n' +
-	'Use this for ad-hoc data exploration and Q&A. ' +
-	'Returns `chatId` and a `url` that opens the chat in the Nao UI — surface the URL to the user as a clickable link so they can jump to the conversation (and any rendered charts/tables). ' +
-	'The returned `chatId` can also be passed to `create_story` as `chat_id` to attach a follow-up story to this conversation.';
+	"Delegate an analytics task to nao's sub-agent — it runs the full reasoning loop " +
+	'(reads project rules/context, writes SQL, builds charts, drafts stories) and the whole ' +
+	'conversation is persisted as a chat visible in the nao UI (replayable, shareable, forkable ' +
+	'by the end user).\n\n' +
+	'USE WHEN: you want to hand off the whole task and produce a user-facing trace in nao — ' +
+	'the thinking process itself matters as much as the final answer.\n' +
+	"SKIP WHEN: you'd rather drive the workflow yourself by chaining `ls` / `grep` / `execute_sql` / " +
+	'`display_chart` / `create_story` step by step — those run as plain tool calls, leave no chat ' +
+	'in the UI, and give you full control over each step.\n\n' +
+	'Returns the assistant text plus `chatId`, `chatUrl`, and arrays `charts` / `stories` produced ' +
+	'during the run. Chain `charts[].queryId` + `chatId` into `display_chart`, or pass `chatId` to ' +
+	'`create_story` to attach a follow-up document. Side effect: creates a chat row.';
 
-export function registerAgentTools(server: McpServer, ctx: McpContext): void {
+export function registerSubAgentTools(server: McpServer, ctx: McpContext): void {
 	server.registerTool(
 		'ask_nao',
 		{
 			title: 'Ask Nao',
 			description: ASK_NAO_DESCRIPTION,
 			inputSchema: {
-				question: z.string().describe('The analytics question.'),
+				question: z
+					.string()
+					.describe(
+						'Natural-language analytics question or task. The agent reads project context ' +
+							'(rules, columns, semantic layer) to decide what to query — no need to mention SQL or table names.',
+					),
 				chatId: z
 					.uuid()
 					.optional()
 					.describe(
 						'UUID of an existing chat to continue. Omit to start a new chat. ' +
-							'Reuse the UUID from a previous ask_nao ONLY when the new question clearly builds on the same data and topic. ' +
-							'If the topic shifts or the prior reply was a refusal, omit it — inheriting that context may cause the follow-up to repeat the refusal.',
+							'Reuse only when the new question clearly builds on the same topic. ' +
+							'If the topic shifts or the prior reply was a refusal, omit it.',
 					),
 			},
 			outputSchema: {
-				chatId: z.uuid().describe('UUID of the chat in which the question was answered.'),
-				chatUrl: z.url().describe('URL to open the chat (and rendered charts/tables) in the Nao UI.'),
+				chatId: z
+					.string()
+					.describe(
+						'UUID of the chat that holds this run. Pass to `display_chart` / `create_story` / `update_story` to attach further work.',
+					),
+				chatUrl: z.url().describe('URL to open the chat in the nao UI.'),
 				text: z.string().describe('The assistant final text response.'),
+				charts: z
+					.array(
+						z.object({
+							queryId: z.string().describe('`query_id` produced by the agent — pass to `display_chart`.'),
+							chartType: z.string(),
+							title: z.string().optional(),
+							block: z.string().describe('`<chart>` block ready to paste into a story `content` field.'),
+						}),
+					)
+					.describe('Charts produced during this run. Chain with `display_chart` using `chatId`.'),
+				stories: z
+					.array(z.object({ id: z.string(), title: z.string() }))
+					.describe('Stories created or updated during this run.'),
 			},
 		},
 		defineMcpHandler(
@@ -53,8 +84,16 @@ export function registerAgentTools(server: McpServer, ctx: McpContext): void {
 				const stream = agent.stream(uiMessages);
 				const text = await consumeStreamWithProgress(stream, extra);
 
-				const output = { chatId: chat.id, chatUrl: chatUrl(chat.id), text };
-				ctx.sessionChatRef.lastChatId = chat.id;
+				const artifacts = agent.generatedArtifacts;
+				const charts = artifacts.charts.map((c) => ({
+					queryId: c.query_id,
+					chartType: c.chart_type,
+					title: c.title,
+					block: buildStoryChartBlock(c),
+				}));
+				const stories = artifacts.stories;
+
+				const output = { chatId: chat.id, chatUrl: chatUrl(chat.id), text, charts, stories };
 				return {
 					content: [
 						{
