+++ apps/backend/src/types/mcp-endpoint.ts
@@ -1,13 +1,11 @@
 export interface McpEndpointSettings {
 	enabled: boolean;
-	agentModeEnabled: boolean;
-	toolsModeEnabled: boolean;
-	objectsModeEnabled: boolean;
+	subAgentModeEnabled: boolean;
+	contextLayerModeEnabled: boolean;
 }
 
 export const DEFAULT_MCP_ENDPOINT_SETTINGS: McpEndpointSettings = {
 	enabled: false,
-	agentModeEnabled: true,
-	toolsModeEnabled: true,
-	objectsModeEnabled: true,
+	subAgentModeEnabled: true,
+	contextLayerModeEnabled: true,
 };
